Add BIP145 segwit support
Transaction "hash" and "txid" now have specific meanings. "txid" represents the
hashed tx without witness data. This is used to form the in-header merkle root
as usual.

"hash" is the tx including witness data, if any. If there is no witness data,
hash == txid.

if the "segwit" rule is active, hashes are combined into a merkle tree, and the
root is inserted as part of the scriptPubKey in a coinbase txout.

See BIP141 for specifics.